change int to long long

+++ src/eval_expr.cpp
@@ -2,22 +2,22 @@
 
 using namespace std;
 
-int CommaExpr:: eval(Scope& scope)
+long long CommaExpr:: eval(Scope& scope)
 {
   return expr1->eval(scope), expr2->eval(scope);
 }
 
-int EqualityExpr:: eval(Scope& scope)
+long long EqualityExpr:: eval(Scope& scope)
 {
-  int val1 = expr1->eval(scope), val2 = expr2->eval(scope);
+  long long val1 = expr1->eval(scope), val2 = expr2->eval(scope);
 
   if ("!=" == op) return val1 != val2;
   return val1 == val2;
 }
 
-int RelationalExpr:: eval(Scope& scope)
+long long RelationalExpr:: eval(Scope& scope)
 {
-  int val1 = expr1->eval(scope), val2 = expr2->eval(scope);
+  long long val1 = expr1->eval(scope), val2 = expr2->eval(scope);
 
   if (">=" == op) return val1 >= val2;
   else if("<=" == op) return val1 <= val2;
@@ -26,36 +26,36 @@ int RelationalExpr:: eval(Scope& scope)
   else return -1;
 }
 
-int AdditiveExpr:: eval(Scope& scope)
+long long AdditiveExpr:: eval(Scope& scope)
 {
-  int val1 = expr1->eval(scope), val2 = expr2->eval(scope);
+  long long val1 = expr1->eval(scope), val2 = expr2->eval(scope);
 
   if ("+" == op) return val1 + val2;
   else if("-" == op) return val1 - val2;
   else return -1;
 }
 
-int MultExpr:: eval(Scope& scope)
+long long MultExpr:: eval(Scope& scope)
 {
-  int val1 = expr1->eval(scope), val2 = expr2->eval(scope);
+  long long val1 = expr1->eval(scope), val2 = expr2->eval(scope);
 
   if ("*" == op) return val1 * val2;
   else if("/" == op) return val1 / val2;
   else return -1;
 }
 
-int UnaryExpr:: eval(Scope& scope)
+long long UnaryExpr:: eval(Scope& scope)
 {
-  int val = expr->eval(scope);
+  long long val = expr->eval(scope);
 
   if ("-" == op) return -val;
   else if("+" == op) return val;
   else return -1;
 }
 
-int PostfixExpr:: eval(Scope& scope)
+long long PostfixExpr:: eval(Scope& scope)
 {
-  int val = expr->eval(scope);
+  long long val = expr->eval(scope);
   string& id = dynamic_pointer_cast<PrimaryExprId>(expr)->id;
 
   if ("++" == op) scope.mod_identifier(id, val + 1);
@@ -64,22 +64,22 @@ int PostfixExpr:: eval(Scope& scope)
   return val;
 }
 
-int PrimaryExprConst:: eval(Scope& scope) {return val;}
+long long PrimaryExprConst:: eval(Scope& scope) {return val;}
 
-int PrimaryExprId:: eval(Scope& scope) {return scope.get_identifier(id);}
+long long PrimaryExprId:: eval(Scope& scope) {return scope.get_identifier(id);}
 
-int AssignmentExpr::eval(Scope& scope)
+long long AssignmentExpr::eval(Scope& scope)
 {
-  int val = expr->eval(scope);
+  long long val = expr->eval(scope);
 
   scope.mod_identifier(id, val);
   
   return val;
 }
 
-int InitDecl::eval(Scope& scope)
+long long InitDecl::eval(Scope& scope)
 {
-  int val = expr->eval(scope);
+  long long val = expr->eval(scope);
 
   scope.set_identifier(id, val);
